+++ graphistry/compute/__init__.py
@@ -2,6 +2,7 @@ from .ComputeMixin import ComputeMixin
 from .ast import (
     n, e_forward, e_reverse, e_undirected
 )
+from .chain import Chain
 from .predicates.is_in import (
     is_in, IsIn
 )